+++ components/FederationMissingNotice.vue
@@ -28,7 +28,10 @@ export default {
 		retry: {
 			type: Function,
 			default: null
-		}
+		},
+		// Mixins don't work properly in web components,
+		// see https://github.com/vuejs/vue-web-component-wrapper/issues/30,
+		...FederationMixin.props
 	},
 	computed: {
 		services() {