+++ components/internal/ProcessParameter.vue
@@ -37,7 +37,10 @@ export default {
 		},
 		processUrl: {
 			type: String
-		}
+		},
+		// Mixins don't work properly in web components,
+		// see https://github.com/vuejs/vue-web-component-wrapper/issues/30,
+		...FederationMixin.props
 	},
 	computed: {
 		hasDefault() {